Huge refactor: Move eventsub broker and create common folder directory
