--- EventSub/README.md
@@ -1,1 +1,0 @@
-**Required version** `1.18.0`
