--- EventSub/melonbot.eventsub.service
@@ -1,17 +1,0 @@
-[Unit]
-Description=EventSub handler for Melonbot
-After=network.target
-
-[Service]
-ExecStart=/my/path/to/the/program/bin/eventsub -config ./../config.json -loglevel info
-WorkingDirectory=/my/path/to/Melonbot/EventSub
-User=UHHHHHHH
-Group=UHHHHHHH
-Restart=always
-RestartSec=5
-StandardOutput=syslog
-StandardError=syslog
-SyslogIdentifier=%n
-
-[Install]
-WantedBy=multi-user.target
