--- EventSub/pkg/util/main.go
@@ -1,9 +1,0 @@
-package util
-
-import log "github.com/sirupsen/logrus"
-
-func Validate(e error) {
-	if e != nil {
-		log.Fatal(e)
-	}
-}
