+++ Golang/EventSub/README.md
@@ -1,0 +1,3 @@
+**Required version** `1.18.0`
+
+**Twitch EventSub Broker**
