+++ Golang/EventSub/go.mod
@@ -1,0 +1,17 @@
+module github.com/JoachimFlottorp/Melonbot/Golang/EventSub
+
+go 1.18
+
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/nicklaw5/helix v1.25.0
+)
+
+require (
+	github.com/cespare/xxhash/v2 v2.1.2 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/golang-jwt/jwt v3.2.2+incompatible // indirect
+	golang.org/x/net v0.0.0-20211015210444-4f30a5c0130f // indirect
+	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e // indirect
+	golang.org/x/text v0.3.7 // indirect
+)
