+++ Golang/go.mod
@@ -1,0 +1,3 @@
+module github.com/JoachimFlottorp/Melonbot/Golang
+
+go 1.18
