+++ package.json
@@ -6,7 +6,7 @@
 	"type": "module",
 	"scripts": {
 		"test": "jest",
-		"test:eventsub": "node Scripts/Build.Go.mjs test",
+		"test:eventsub": "make -C Golang/EventSub test",
 		"start:bot": "npm run build:tsc && node --inspect build/bin/master.js",
 		"start:web": "node --inspect build/bin/web.js",
 		"setup": "npm install && npm run build && node build/bin/setup.js",
